Moved tests, added missing attributes to mock v1 classes
